Move stories to stories/ folder
