+++ packages/components/src/calendar/stories/date-calendar.stories.tsx
@@ -19,7 +19,7 @@ import {
 	sv,
 } from 'date-fns/locale';
 import { useState, useEffect } from 'react';
-import { DateCalendar, TZDate } from './index';
+import { DateCalendar, TZDate } from '../index';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta: Meta< typeof DateCalendar > = {
